fix: 修正 GetCursorPos 的 POINT 结构声明（double→int）导致位置恒为左上角
根因：POINT 应为两个 32 位 int（共 8 字节），误声明为 double（16 字节），
P/Invoke 返回失败，工具栏坐标退化为 (0,0) 而固定在屏幕左上角。
已用真实光标坐标验证 int 定义可正确返回 (747,1290) 非零坐标。

+++ src/FloatingHelper.App/App.xaml.cs
@@ -112,8 +112,8 @@ public partial class App : System.Windows.Application
     [StructLayout(LayoutKind.Sequential)]
     private struct POINT
     {
-        public double X;
-        public double Y;
+        public int X;
+        public int Y;
     }
 
     [DllImport("user32.dll")]
